introduces Add, Delete and Update Activity functionality

Closes #60 and #61

See merge request cse1105/2021-2022/team-repositories/oopp-group-27/repository-template!75

+++ commons/src/main/java/commons/Activity.java
@@ -102,7 +102,7 @@ public class Activity implements Reachable{
      * @return consumption of the activity
      */
 
-    public double getConsumption() {
+    public long getConsumption() {
         return consumption;
     }
